+++ modules/meteroid/src/api/services/subscriptions/service.rs
@@ -275,7 +275,7 @@ impl SubscriptionsService for SubscriptionServiceComponents {
             currency: subscription.currency.clone(),
             days_until_due: subscription.net_terms,
             line_items: serialized_invoice_lines,
-            amount_cents: Some(invoice_lines.total as i32),
+            amount_cents: Some(invoice_lines.total),
         };
 
         db::invoices::create_invoice()
@@ -479,7 +479,7 @@ impl SubscriptionsService for SubscriptionServiceComponents {
                 })?;
 
             let invoice_lines = vec![invoice_line];
-            let total = Some(invoice_lines.iter().map(|line| line.total as i32).sum());
+            let total = Some(invoice_lines.iter().map(|line| line.total).sum());
 
             let serialized_invoice_lines = serde_json::to_value(invoice_lines).map_err(|e| {
                 Status::internal("Failed to serialize invoice lines")
